pyproject.toml
[tool.poetry]
name = "pmid2fulltext"
version = "0.1.0"
description = ""
authors = ["iQuxLE <ckroll95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
biopython = "^1.84"
click = "^8.1.7"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

